scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "openai>=2.26.0",
    "python-dotenv>=1.0.0",
]
